# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Website Builder Platform API — bilingual (AR/EN)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    get:
      operationId: authLoginRedirect
      tags: [auth]
      summary: Start login flow (Replit)
      description: Redirects to Replit OIDC login (replit provider only)
      responses:
        "302":
          description: Redirect to Replit Auth
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login with email and password
      description: Authenticates a user with email and password (local provider only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
              required:
                - email
                - password
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/callback:
    get:
      operationId: authCallback
      tags: [auth]
      summary: OIDC callback
      description: Handles Replit OIDC callback, creates session, redirects to app
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to app after successful auth
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register with email and password
      description: Creates a new user account (local provider only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
              required:
                - email
                - password
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      description: Returns the authenticated user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update profile
      description: Update display name or locale preference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            message_ar:
              type: string
          required:
            - code
            - message
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        locale:
          type: string
          enum: [ar, en]
        dailyLimitUsd:
          type: number
        monthlyLimitUsd:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - locale
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        locale:
          type: string
          enum: [ar, en]
...